CreateImageWizard: Add DocumentationButton form component
A documentation button, which contains an external link to the
official documentation, appears several times throughout the frontend.
Currently, the code for the documentation button is copy/pasted where
needed. Adding a form component for the documentation button allows code reuse
and improves readability in the code that calls it.

Because this component is used by the LandingPage, ImagesTable, and
ImageWizard it has been placed in a new directory,
`src/Components/sharedComponents` which is at the same level as these
directories.