.github: Sync branches workflow
This workflow syncs the main branch with the stage-stable branch.

It also includes a production target which needs to be dispatched
manually. The production target pushes the input ref to both prod-beta
and prod-stable branches.

This results in the main branch always being deployed on stage-beta and
stage-stable, and the prod-beta and prod-stable branches being synced on
release.