devel: update composer config
Update some of the config for the docker-compose
file since there have been a few changes and the
config is slightly out of date. Additionally, a
refresh_token parameter was added to the oauth_url
since this is used by composer's mock openid provider
to generate the `rh-id-org` key in the jwt token.