Wizard: Fix sources select bug
Fixes #1007. A bug in Patternfly causes dropdown and select menus to
appear underneath the Wizard footer when they are appended to their
parent (which is done to allow them to 'pop out' of the wizard and avoid
scrolling inside of the wizard's inner window).

The bug is discussed here:
https://github.com/patternfly/patternfly/issues/4794#issuecomment-1451187628

For now, the best solution seems to be adjusting the wizard's z-index
values as the other solutions discussed have a negative impact on
accessibility and keyboard navigation.