Typescript: Add types for API schema
This commit adds types for the APIs that we currently query using RTK
Query. The OpenAPI schema were converted to Typescript types using the
openapi-typescript package: https://www.npmjs.com/package/openapi-typescript

In the future, we would like to add a script to a /tools directory that
will automatically generate the types from the schema. That isn't
possible now because the OpenAPI schema on consoledot require
authorization to acccess, but exposing them without auth is on the
roadmap and once that is complete we will add a script to automate the
process of generating these types.